+++ packages/app/src/sandbox/eval/transpiled-module.ts
@@ -8,7 +8,7 @@ import hashsum from 'hash-sum';
 
 import * as pathUtils from '@codesandbox/common/lib/utils/path';
 
-import { Module } from './entities/module';
+import { Module } from './types/module';
 import { SourceMap } from './transpilers/utils/get-source-map';
 import ModuleError from './errors/module-error';
 import ModuleWarning from './errors/module-warning';
